refactor: move metadata provider classes to analytics package
