+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics/domain_service/PlanMetadataProvider.java
@@ -13,37 +13,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.apim.core.api.domain_service;
+package io.gravitee.apim.core.analytics.domain_service;
 
 import io.gravitee.apim.core.DomainService;
-import io.gravitee.apim.core.analytics.domain_service.AnalyticsMetadataProvider;
-import io.gravitee.apim.core.api.crud_service.ApiCrudService;
-import io.gravitee.apim.core.api.exception.ApiNotFoundException;
-import io.gravitee.apim.core.api.model.Api;
+import io.gravitee.apim.core.plan.crud_service.PlanCrudService;
+import io.gravitee.apim.core.plan.exception.PlanNotFoundException;
 import java.util.HashMap;
 import java.util.Map;
 
 @DomainService
-public class ApiMetadataProvider implements AnalyticsMetadataProvider {
+public class PlanMetadataProvider implements AnalyticsMetadataProvider {
 
     private static final String METADATA_NAME = "name";
     private static final String METADATA_DELETED = "deleted";
     private static final String METADATA_UNKNOWN = "unknown";
-    private static final String METADATA_VERSION = "version";
-    private static final String METADATA_UNKNOWN_API_NAME = "Unknown API (not found)";
-    private static final String METADATA_DELETED_API_NAME = "Deleted API";
+    private static final String METADATA_UNKNOWN_PLAN_NAME = "Unknown plan (keyless)";
+    private static final String METADATA_DELETED_PLAN_NAME = "Deleted plan";
     private static final String UNKNOWN_SERVICE = "1";
     private static final String UNKNOWN_SERVICE_MAPPED = "?";
 
-    private final ApiCrudService apiCrudService;
+    private final PlanCrudService planCrudService;
 
-    public ApiMetadataProvider(ApiCrudService apiCrudService) {
-        this.apiCrudService = apiCrudService;
+    public PlanMetadataProvider(PlanCrudService planCrudService) {
+        this.planCrudService = planCrudService;
     }
 
     @Override
     public boolean appliesTo(Field field) {
-        return field == Field.API;
+        return field == Field.PLAN;
     }
 
     @Override
@@ -51,19 +48,15 @@ public class ApiMetadataProvider implements AnalyticsMetadataProvider {
         Map<String, String> metadata = new HashMap<>();
         try {
             if (UNKNOWN_SERVICE.equals(key) || UNKNOWN_SERVICE_MAPPED.equals(key)) {
-                metadata.put(METADATA_NAME, METADATA_UNKNOWN_API_NAME);
+                metadata.put(METADATA_NAME, METADATA_UNKNOWN_PLAN_NAME);
                 metadata.put(METADATA_UNKNOWN, Boolean.TRUE.toString());
             } else {
-                var api = apiCrudService.get(key);
-                metadata.put(METADATA_NAME, api.getName());
-                metadata.put(METADATA_VERSION, api.getVersion());
-                if (Api.ApiLifecycleState.ARCHIVED == api.getApiLifecycleState()) {
-                    metadata.put(METADATA_DELETED, Boolean.TRUE.toString());
-                }
+                var plan = planCrudService.getById(key);
+                metadata.put(METADATA_NAME, plan.getName());
             }
-        } catch (ApiNotFoundException e) {
+        } catch (PlanNotFoundException e) {
             metadata.put(METADATA_DELETED, Boolean.TRUE.toString());
-            metadata.put(METADATA_NAME, METADATA_DELETED_API_NAME);
+            metadata.put(METADATA_NAME, METADATA_DELETED_PLAN_NAME);
         }
         return metadata;
     }
